add combined worker engine to use parallel and sequence simultaneously
