+++ handler/dummy/cooler.go
@@ -1,0 +1,7 @@
+package dummy
+
+import "time"
+
+func (d *Dummy) Cooler() time.Duration {
+	return 10 * time.Second
+}
