+++ handler/dummy/dummy.go
@@ -1,0 +1,7 @@
+package dummy
+
+type Dummy struct{}
+
+func New() *Dummy {
+	return &Dummy{}
+}
