+++ handler/dummy/ensure.go
@@ -1,0 +1,5 @@
+package dummy
+
+func (d *Dummy) Ensure() error {
+	return nil
+}
