+++ worker/parallel/daemon_test.go
